+++ Microsoft.Azure.Cosmos.Encryption/src/EncryptionContainer.cs
@@ -54,7 +54,8 @@ namespace Microsoft.Azure.Cosmos.Encryption
                 throw new ArgumentNullException(nameof(item));
             }
 
-            if (requestOptions is EncryptionItemRequestOptions encryptionItemRequestOptions)
+            if (requestOptions is EncryptionItemRequestOptions encryptionItemRequestOptions &&
+                encryptionItemRequestOptions.EncryptionOptions != null)
             {
                 if (partitionKey == null)
                 {
@@ -97,7 +98,8 @@ namespace Microsoft.Azure.Cosmos.Encryption
             CosmosDiagnosticsContext diagnosticsContext = CosmosDiagnosticsContext.Create(requestOptions);
             using (diagnosticsContext.CreateScope("CreateItemStream"))
             {
-                if (requestOptions is EncryptionItemRequestOptions encryptionItemRequestOptions)
+                if (requestOptions is EncryptionItemRequestOptions encryptionItemRequestOptions &&
+                    encryptionItemRequestOptions.EncryptionOptions != null)
                 {
                     streamPayload = await EncryptionProcessor.EncryptAsync(
                         streamPayload,
@@ -221,7 +223,8 @@ namespace Microsoft.Azure.Cosmos.Encryption
                 throw new ArgumentNullException(nameof(item));
             }
 
-            if (requestOptions is EncryptionItemRequestOptions encryptionItemRequestOptions)
+            if (requestOptions is EncryptionItemRequestOptions encryptionItemRequestOptions &&
+                encryptionItemRequestOptions.EncryptionOptions != null)
             {
                 if (partitionKey == null)
                 {
@@ -272,7 +275,8 @@ namespace Microsoft.Azure.Cosmos.Encryption
             CosmosDiagnosticsContext diagnosticsContext = CosmosDiagnosticsContext.Create(requestOptions);
             using (diagnosticsContext.CreateScope("ReplaceItemStream"))
             {
-                if (requestOptions is EncryptionItemRequestOptions encryptionItemRequestOptions)
+                if (requestOptions is EncryptionItemRequestOptions encryptionItemRequestOptions &&
+                    encryptionItemRequestOptions.EncryptionOptions != null)
                 {
                     if (partitionKey == null)
                     {
@@ -322,7 +326,8 @@ namespace Microsoft.Azure.Cosmos.Encryption
                 throw new ArgumentNullException(nameof(item));
             }
 
-            if (requestOptions is EncryptionItemRequestOptions encryptionItemRequestOptions)
+            if (requestOptions is EncryptionItemRequestOptions encryptionItemRequestOptions &&
+                encryptionItemRequestOptions.EncryptionOptions != null)
             {
                 if (partitionKey == null)
                 {
@@ -363,7 +368,8 @@ namespace Microsoft.Azure.Cosmos.Encryption
             CosmosDiagnosticsContext diagnosticsContext = CosmosDiagnosticsContext.Create(requestOptions);
             using (diagnosticsContext.CreateScope("UpsertItemStream"))
             {
-                if (requestOptions is EncryptionItemRequestOptions encryptionItemRequestOptions)
+                if (requestOptions is EncryptionItemRequestOptions encryptionItemRequestOptions &&
+                    encryptionItemRequestOptions.EncryptionOptions != null)
                 {
                     if (partitionKey == null)
                     {
@@ -404,7 +410,10 @@ namespace Microsoft.Azure.Cosmos.Encryption
         public override TransactionalBatch CreateTransactionalBatch(
             PartitionKey partitionKey)
         {
-            return this.container.CreateTransactionalBatch(partitionKey);
+            return new EncryptionTransactionalBatch(
+                this.container.CreateTransactionalBatch(partitionKey),
+                this.encryptor,
+                this.cosmosSerializer);
         }
 
         public override Task<ContainerResponse> DeleteContainerAsync(