+++ Microsoft.Azure.Cosmos.Encryption/src/EncryptionProcessor.cs
@@ -18,6 +18,11 @@ namespace Microsoft.Azure.Cosmos.Encryption
     {
         internal static readonly CosmosJsonDotNetSerializer baseSerializer = new CosmosJsonDotNetSerializer();
 
+        /// <remarks>
+        /// If there isn't any PathsToEncrypt, input stream will be returned without any modification.
+        /// Else input stream will be disposed, and a new stream is returned.
+        /// In case of an exception, input stream won't be disposed, but position will be end of stream.
+        /// </remarks>
         public static async Task<Stream> EncryptAsync(
             Stream input,
             Encryptor encryptor,
@@ -25,10 +30,22 @@ namespace Microsoft.Azure.Cosmos.Encryption
             CosmosDiagnosticsContext diagnosticsContext,
             CancellationToken cancellationToken)
         {
-            Debug.Assert(input != null);
-            Debug.Assert(encryptor != null);
-            Debug.Assert(encryptionOptions != null);
             Debug.Assert(diagnosticsContext != null);
+
+            if (input == null)
+            {
+                throw new ArgumentNullException(nameof(input));
+            }
+
+            if (encryptor == null)
+            {
+                throw new ArgumentNullException(nameof(encryptor));
+            }
+
+            if (encryptionOptions == null)
+            {
+                throw new ArgumentNullException(nameof(encryptionOptions));
+            }
 
             if (string.IsNullOrWhiteSpace(encryptionOptions.DataEncryptionKeyId))
             {
@@ -64,15 +81,14 @@ namespace Microsoft.Azure.Cosmos.Encryption
 
             foreach (string pathToEncrypt in encryptionOptions.PathsToEncrypt)
             {
-                string propertyName = pathToEncrypt.Substring(1);
-                JToken propertyValueHolder = itemJObj.Property(propertyName).Value;
-
-                // Even null in the JSON is a JToken with Type Null, this null check is just a sanity check
-                if (propertyValueHolder != null)
+                string propertyName = pathToEncrypt.Substring(1);                
+                if (!itemJObj.TryGetValue(propertyName, out JToken propertyValue))
                 {
-                    toEncryptJObj.Add(propertyName, propertyValueHolder.Value<JToken>());
-                    itemJObj.Remove(propertyName);
+                    throw new ArgumentException($"{nameof(encryptionOptions.PathsToEncrypt)} includes a path: '{pathToEncrypt}' which was not found.");
                 }
+
+                toEncryptJObj.Add(propertyName, propertyValue.Value<JToken>());
+                itemJObj.Remove(propertyName);
             }
 
             MemoryStream memoryStream = EncryptionProcessor.baseSerializer.ToStream<JObject>(toEncryptJObj);
@@ -98,9 +114,15 @@ namespace Microsoft.Azure.Cosmos.Encryption
                 encryptedData: cipherText);
 
             itemJObj.Add(Constants.EncryptedInfo, JObject.FromObject(encryptionProperties));
+            input.Dispose();
             return EncryptionProcessor.baseSerializer.ToStream(itemJObj);
         }
 
+        /// <remarks>
+        /// If there isn't any data that needs to be decrypted, input stream will be returned without any modification.
+        /// Else input stream will be disposed, and a new stream is returned.
+        /// In case of an exception, input stream won't be disposed, but position will be end of stream.
+        /// </remarks>
         public static async Task<Stream> DecryptAsync(
             Stream input,
             Encryptor encryptor,
@@ -148,6 +170,7 @@ namespace Microsoft.Azure.Cosmos.Encryption
             }
 
             itemJObj.Remove(Constants.EncryptedInfo);
+            input.Dispose();
             return EncryptionProcessor.baseSerializer.ToStream(itemJObj);
         }