+++ Microsoft.Azure.Cosmos.Samples/Tools/Benchmark/Fx/IExecutionStrategy.cs
@@ -1,0 +1,24 @@
+﻿//------------------------------------------------------------
+// Copyright (c) Microsoft Corporation.  All rights reserved.
+//------------------------------------------------------------
+
+namespace CosmosBenchmark
+{
+    using System.Threading.Tasks;
+
+    internal interface IExecutionStrategy
+    {
+        public static IExecutionStrategy StartNew(
+            BenchmarkConfig config,
+            IBenchmarkOperatrion benchmarkOperation)
+        {
+            return new ParallelExecutionStrategy(benchmarkOperation);
+        }
+
+        public Task ExecuteAsync(
+            int serialExecutorConcurrency,
+            int serialExecutorIterationCount,
+            double warmupFraction);
+
+    }
+}