+++ Microsoft.Azure.Cosmos/src/EnvironmentInformation.cs
@@ -9,11 +9,14 @@ namespace Microsoft.Azure.Cosmos
 
     internal sealed class EnvironmentInformation
     {
+        private const int MaxClientId = 10;
         private static readonly string clientSDKVersion;
         private static readonly string directPackageVersion;
         private static readonly string framework;
         private static readonly string architecture;
         private static readonly string os;
+        private static readonly object clientCountLock = new object();
+        private static int clientId = 0;
 
         static EnvironmentInformation()
         {
@@ -28,8 +31,16 @@ namespace Microsoft.Azure.Cosmos
 
         public EnvironmentInformation()
         {
-            string now = DateTime.UtcNow.Ticks.ToString();
-            this.ClientId = now.Substring(now.Length - 5); // 5 most significative digits
+            lock (EnvironmentInformation.clientCountLock)
+            {
+                int newClientId = EnvironmentInformation.MaxClientId;
+                if (EnvironmentInformation.clientId <= EnvironmentInformation.MaxClientId)
+                {
+                    newClientId = EnvironmentInformation.clientId++;
+                }
+
+                this.ClientId = newClientId.ToString().PadLeft(2, '0');
+            }
         }
 
         /// <summary>
@@ -64,5 +75,13 @@ namespace Microsoft.Azure.Cosmos
         /// </summary>
         /// <seealso cref="RuntimeInformation.ProcessArchitecture"/>
         public string ProcessArchitecture => EnvironmentInformation.architecture;
+
+        /// <summary>
+        /// Only used to reset counter on tests.
+        /// </summary>
+        public static void ResetCounter()
+        {
+            EnvironmentInformation.clientId = 0;
+        }
     }
 }