+++ Microsoft.Azure.Cosmos/src/Resource/CosmosExceptions/CosmosException.cs
@@ -236,6 +236,7 @@ namespace Microsoft.Azure.Cosmos
 
             if (this.Diagnostics != null)
             {
+                stringBuilder.Append("--- Cosmos Diagnostics ---");
                 stringBuilder.Append(this.Diagnostics);
             }